Rename distro directory to distro_lib.
This prevents conflicts with the Python 3 distro module.
