+++ google_compute_engine/distro_lib/debian_9/utils.py
@@ -13,14 +13,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Utilities that are distro specific for use on EL 6."""
+"""Utilities that are distro specific for use on Debian 9."""
 
-from google_compute_engine.distro import helpers
-from google_compute_engine.distro import utils
+from google_compute_engine.distro_lib import helpers
+from google_compute_engine.distro_lib import utils
 
 
 class Utils(utils.Utils):
-  """Utilities used by Linux guest services on Debian 8."""
+  """Utilities used by Linux guest services on Debian 9."""
 
   def EnableNetworkInterfaces(
       self, interfaces, logger, dhclient_script=None):
@@ -31,7 +31,7 @@ class Utils(utils.Utils):
       logger: logger object, used to write to SysLog and serial port.
       dhclient_script: string, the path to a dhclient script used by dhclient.
     """
-    helpers.CallDhclient(interfaces, logger, dhclient_script=dhclient_script)
+    helpers.CallDhclient(interfaces, logger)
 
   def HandleClockSync(self, logger):
     """Sync the software clock with the hypervisor clock.
