+++ google_compute_engine/distro_lib/freebsd_11/tests/utils_test.py
@@ -15,7 +15,7 @@
 
 """Unittest for utils.py module."""
 
-from google_compute_engine.distro.debian_8 import utils
+from google_compute_engine.distro_lib.freebsd_11 import utils
 from google_compute_engine.test_compat import mock
 from google_compute_engine.test_compat import unittest
 
@@ -26,7 +26,7 @@ class UtilsTest(unittest.TestCase):
     self.mock_logger = mock.Mock()
     self.mock_setup = mock.create_autospec(utils.Utils)
 
-  @mock.patch('google_compute_engine.distro.helpers.CallDhclient')
+  @mock.patch('google_compute_engine.distro_lib.helpers.CallDhclient')
   def testEnableNetworkInterfaces(self, mock_call):
     mocks = mock.Mock()
     mocks.attach_mock(mock_call, 'call')
@@ -36,7 +36,7 @@ class UtilsTest(unittest.TestCase):
     expected_calls = [mock.call.call(['A', 'B'], mock.ANY)]
     self.assertEqual(mocks.mock_calls, expected_calls)
 
-  @mock.patch('google_compute_engine.distro.helpers.CallHwclock')
+  @mock.patch('google_compute_engine.distro_lib.helpers.CallNtpdate')
   def testHandleClockSync(self, mock_call):
     mocks = mock.Mock()
     mocks.attach_mock(mock_call, 'call')
