+++ google_compute_engine/distro_lib/helpers.py
