+++ google_compute_engine/distro_lib/utils.py
